refactor: OpenButtonのts化
